+++ src/main/java/br/edu/tglima/locadora/models/veiculo/OpStatus.java
@@ -1,0 +1,17 @@
+package br.edu.tglima.locadora.models.veiculo;
+
+public enum OpStatus {
+
+	ALUGADO("Alugado"), DISPONIVEL("Disponível"), INOPERANTE("Inoperante"), MANUTENCAO("Manutenção");
+
+	private String label;
+
+	private OpStatus(String label) {
+		this.label = label;
+	}
+
+	public String getLabel() {
+		return label;
+	}
+
+}